Reduce constant dups for identifiers w/ hash table
When accessing a global, we'll try and load the constant from the
some sort of global state for the current chunk.

Before: identifier access adds a constant to the chunk.
After: identifiers are deduped by name, and the mapping between
string->constant is kept in a table.

The tradeoff I think is:
- More memory per unique global (1 entry in the table)
- Constant memory usage regardless of the number of references to
  the global
- Slower at compile time