Use hash table to make var lookup faster
Honestly, this is kinda gross. Var lookup might be faster because
we do not need to linearly scan, but there's a hash lookup which is
not free at all, and on every `endScope` we loop through _all_ locals
to evict the ones that need to go out of scope. It's likely slower,
all in all.

Meh.