Remove the need to unmark obj after GC
By using a "gc mark" that changes between GC runs, we don't
need to reset the mark on each object after a GC. I'm using an int
but anything would do, heck we could even keep this as a bool, but
flip it between GC run, we'd still get the same result.

On a very synthetic Lox script with GC stress on, this version is 5%
faster.